CLI: move DryRunExecutors to separate module
